_cli, _utils: Move split chain helper to utilities module
